Implement constituent globs
Given that CppNix' n-e-j already uses my patches to bring back
constituents[1], I decided to also submit the part that allows globbing
in hydra-eval-jobs to this repo[2].

This basically allows you to do

    constituents = [ "nixos.*" ];

to select all jobs starting with nixos. as constituents.
Additionally, it fixes a bug in Hydra (that's also present here probably) where
transitive constituent jobs may be discarded silently if declared by string
instead of by derivation.

Given that the code became slightly larger and this part of the
evaluation shouldn't blow up `main` even more, I decided to move it to
its own compilation unit. Also fixed all the suggestions from the linter
in the diff.

[1] https://determinate.systems/posts/hydra-deployment-source-of-truth/#aggregate-jobs
[2] https://github.com/NixOS/hydra/pull/1425